Add AbstractRouter
